+++ cmd/containerd-shim/main_unix.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"net"
 	"os"
-	"os/signal"
 	"strings"
 
 	"golang.org/x/sys/unix"
@@ -17,9 +16,7 @@ import (
 	shimapi "github.com/containerd/containerd/api/services/shim"
 	"github.com/containerd/containerd/linux/shim"
 	"github.com/containerd/containerd/reaper"
-	"github.com/containerd/containerd/sys"
 	"github.com/containerd/containerd/version"
-	runc "github.com/containerd/go-runc"
 	"github.com/urfave/cli"
 )
 
@@ -81,21 +78,6 @@ func main() {
 	}
 }
 
-// setupSignals creates a new signal handler for all signals and sets the shim as a
-// sub-reaper so that the container processes are reparented
-func setupSignals() (chan os.Signal, error) {
-	signals := make(chan os.Signal, 2048)
-	signal.Notify(signals)
-	// make sure runc is setup to use the monitor
-	// for waiting on processes
-	runc.Monitor = reaper.Default
-	// set the shim as the subreaper for all orphaned processes created by the container
-	if err := sys.SetSubreaper(1); err != nil {
-		return nil, err
-	}
-	return signals, nil
-}
-
 // serve serves the grpc API over a unix socket at the provided path
 // this function does not block
 func serve(server *grpc.Server, path string) error {
@@ -130,9 +112,4 @@ func handleSignals(signals chan os.Signal, server *grpc.Server) error {
 		}
 	}
 	return nil
-}
-
-// setupRoot sets up the root as the shim is started in its own mount namespace
-func setupRoot() error {
-	return unix.Mount("", "/", "", unix.MS_SLAVE|unix.MS_REC, "")
 }
