+++ vendor.conf
@@ -1,5 +1,5 @@
 github.com/coreos/go-systemd 48702e0da86bd25e76cfef347e2adeb434a0d0a6
-github.com/containerd/go-runc 5fe4d8cb7fdc0fae5f5a7f4f1d65a565032401b2
+github.com/containerd/go-runc 66d084de463f26efe84e6a5669e573fa6202e814
 github.com/containerd/console a3863895279f5104533fd999c1babf80faffd98c
 github.com/containerd/cgroups 7b2d1a0f50963678d5799e29d17a4d611f5a5dee
 github.com/docker/go-metrics 8fd5772bf1584597834c6f7961a530f06cbfbb87
